feat: add initial agents management
